Move in-game menu into separate class
