+++ src/CMakeLists.txt
@@ -268,6 +268,8 @@ set(core_sources
     ui/hud_renderer.hpp
     ui/imgui_integration.cpp
     ui/imgui_integration.hpp
+    ui/ingame_menu.cpp
+    ui/ingame_menu.hpp
     ui/ingame_message_display.cpp
     ui/ingame_message_display.hpp
     ui/intro_movie.cpp
